+++ README.md
@@ -1,38 +1,16 @@
-This is a [Next.js](https://nextjs.org/) project bootstrapped with [`create-next-app`](https://github.com/vercel/next.js/tree/canary/packages/create-next-app).
+# Pokemon-Card-Generator 2.0
+This is currently in the process of being refactored; for the more up to date code that hasn't been released yet, check out the `generator-v2` branch. This was originally my first ever project using an API, first built in week 2 of School of Code, now being rebuild to be a scalable, much more interesting React project using NextJS.
 
-## Getting Started
+See the project live: https://piperbates.github.io/pokemon-card-generator/ (Currently showing the older, vanilla JS version)
 
-First, run the development server:
+Makes use of PokeAPI.co to randomly generate a Pokemon card. You can also choose your own Pokemon by its dex number, and cycle through the Pokedex using the arrow keys.
 
-```bash
-npm run dev
-# or
-yarn dev
-# or
-pnpm dev
-```
+## How to use the Pokemon card generator.
+Running it locally: `npm run dev`.
 
-Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
+# Current goals
+More scalable approach
+Some testing
+Validation in case the API is down or is updated in some way
 
-You can start editing the page by modifying `pages/index.tsx`. The page auto-updates as you edit the file.
-
-[API routes](https://nextjs.org/docs/api-routes/introduction) can be accessed on [http://localhost:3000/api/hello](http://localhost:3000/api/hello). This endpoint can be edited in `pages/api/hello.ts`.
-
-The `pages/api` directory is mapped to `/api/*`. Files in this directory are treated as [API routes](https://nextjs.org/docs/api-routes/introduction) instead of React pages.
-
-This project uses [`next/font`](https://nextjs.org/docs/basic-features/font-optimization) to automatically optimize and load Inter, a custom Google Font.
-
-## Learn More
-
-To learn more about Next.js, take a look at the following resources:
-
-- [Next.js Documentation](https://nextjs.org/docs) - learn about Next.js features and API.
-- [Learn Next.js](https://nextjs.org/learn) - an interactive Next.js tutorial.
-
-You can check out [the Next.js GitHub repository](https://github.com/vercel/next.js/) - your feedback and contributions are welcome!
-
-## Deploy on Vercel
-
-The easiest way to deploy your Next.js app is to use the [Vercel Platform](https://vercel.com/new?utm_medium=default-template&filter=next.js&utm_source=create-next-app&utm_campaign=create-next-app-readme) from the creators of Next.js.
-
-Check out our [Next.js deployment documentation](https://nextjs.org/docs/deployment) for more details.
+*** Thanks so much for checking out my project! Find me on LinkedIn: http://www.linkedin.com/in/piperbates ***
