+++ src/styles/PageLayout.module.css
@@ -1,10 +1,11 @@
 .pagelayout {
-    width: 500px;
+    width: 600px;
     margin: auto;
 }
 
 .pagelayout h1 {
     margin-bottom: 10px;
+    text-align: center;
 }
 
 
@@ -33,3 +34,11 @@
 .footer a:hover {
     text-decoration: underline;
 }
+
+
+
+@media only screen and (max-width: 600px)  {
+    .pagelayout {
+        width: 400px;
+    }
+  }
